Adding tab menu: Adding each tab's layout file. Adding each tab java class. Adding page adapter to swipe views between tabs.
